+++ src/libgpaste/gpaste-daemon/gpaste-clipboard-provider.c
@@ -12,7 +12,7 @@ enum
     LAST_SIGNAL
 };
 
-static guint64 signals[LAST_SIGNAL] = { 0 };
+static guint signals[LAST_SIGNAL] = { 0 };
 
 static void
 g_paste_clipboard_provider_default_init (GPasteClipboardProviderInterface *iface G_GNUC_UNUSED)
